Auto MPG Data Set

Predicting city-cycle fuel consumption (miles per gallon) from automobile characteristics.

auto_mpg

Format

A data frame with 388 rows and 9 variables:

mpg

numeric; miles per gallon.

cyl

numeric; number of cylinders.

disp

numeric; engine displacement.

hp

numeric; horse power.

wt

numeric; weight.

accel

numeric; acceleration.

year

numeric; make year.

origin

factor; origin.

Source

ICU Machine Learning Repository. This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University. The dataset was used in the 1983 American Statistical Association Exposition.

Note

Row names are the automobile names with appended year. Four cases with duplicate row names were deleted. Six cases with missing values were also deleted.