Lift and gain charts

Create lift and gain charts for binary classification problems

ggliftPlot(actual, prob, positive = "yes")

Arguments

actual

actual class.

prob

predicted probability of target class.

positive

which class is the target class?

Value

a data frame with lift and gain values for each decile.

Examples

# logistic regression example
fit <- glm(am ~ . , mtcars, family=binomial)
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
pred <- predict(fit, newdata=mtcars, type="response")
results <- ggliftPlot(mtcars$am, pred, positive="1")
#>           0%          10%          20%          30%          40%          50% 
#> 2.220446e-16 2.220446e-16 2.220446e-16 3.059812e-14 5.681889e-12 2.060197e-11 
#>          60%          70%          80%          90%         100% 
#> 6.000000e-01 1.000000e+00 1.000000e+00 1.000000e+00 1.000000e+00 
#> Error in cut.default(prob, breaks = quant10, include.lowest = TRUE, labels = seq(1:10)): 'breaks' are not unique
print(results)
#> Error in print(results): object 'results' not found