Model summary for a "glm" object

Summarize a logistic regression model of class "glm".

# S3 method for glm
info(x)

Arguments

x

an object of class "glm"

Value

A list of class c("info.glm", "list").

overall

omnibus.test

fit.indices

anova.table

coefficient.table

odds.ratios

Details

The function info.glm produces a summary for a logistic regression model fitted with the glm or regress function. There are six sections.

Overall

Model formula, data frame, and sample size (N), and predicted category.

Omnibus Test

Chi-square statistic, degrees of freedom, p-value.

Fit Measures

Stukel's GOF Test, Tjur's psuedo R squared, and Akaike's information criterion (AIC).

Analysis of Variance

ANOVA table with type III (marginal) effects.

Regression Coefficients

Regression coefficients (B), standard errors (SE), z-values, and p-values.

Odds Ratios

odds ratios with 95% confidence intervals

The ANOVA table is obtained from the Anova function in the car package. Odds ratios and confidence intervals are obtained from the Confint function from the car package.

Tjur's pseudo-R.squared (2009) is defined as the mean absolute difference between the mean predicted probability for the positive group and the mean predicted probability for the negative group. It ranges from zero to one.

Stukel provides a goodness of fit (GOF) fit test for ungrouped data, and has better properties than the traditional Hosmer-Lemeshow test in simulation studies. A significant result suggests model misspecification. See Allison (2014) for details.

References

Allison, P. D. (2014). "Measures of fit for logistic regression", SAS Global Forum https://statisticalhorizons.com/wp-content/uploads/GOFForLogisticRegression-Paper.pdf

Allison, P. D. (2014). "Another goodness-of-fit test for logistic regression". https://statisticalhorizons.com/another-goodness-of-fit-test-for-logistic-regression

Stukel, T. A. (1988). “Generalized Logistic Models.” Journal of the American Statistical Association 83:426–431.

Tjur, T. (2009) “Coefficients of determination in logistic regression models—A new proposal: The coefficient of discrimination.” The American Statistician 63: 366-372.

See also

Examples

fit <- regress(am ~ hp + wt + mpg, mtcars)
info(fit)
#> LOGISTIC REGRESSION SUMMARY 
#> Formula: am ~ hp + wt + mpg
#> Data   : mtcars
#> N      : 32
#> 
#> Predicted category: 1
#> 
#> Omnibus Test 
#> Chi-square(3) = 34.4636, p = 1.581e-07 ***
#> 
#> Fit Measures 
#> Stukel's GOF Test: Chi-square(2) = 8.7661, p < 0.0125
#> Tjura's Psuedo-R.squared: 0.8216
#> AIC: 16.7661 
#> 
#> Anova Table (type III tests) 
#>     LR Chisq DF Pr(>Chisq)   
#> hp     8.418  1   0.003715 **
#> wt    10.466  1   0.001216 **
#> mpg    1.293  1   0.255496   
#> 
#> Logistic Regression Coefficients 
#>                     B       SE       z Pr(>|z|)  
#> (Intercept) -15.72137 40.00281 -0.3930  0.69431  
#> hp            0.08389  0.08228  1.0196  0.30790  
#> wt           -6.95492  3.35297 -2.0743  0.03806 *
#> mpg           1.22930  1.58109  0.7775  0.43686  
#> 
#> Odds Ratios (with 95% Confidence Intervals) 
#>     Odds Ratio      2.5%     97.5%
#> hp   1.0875130 1.014e+00    1.5532
#> wt   0.0009539 1.002e-08    0.1163
#> mpg  3.4188427 6.195e-01 1710.0900