Summarise an 'mreg' object

This function serves as a summary method for class 'mreg'. It computes summary statistics and results for a fitted linear model.

# S3 method for mreg
summary(x)

Arguments

x
an object of class 'mreg', presumably an output from the mreg() function.

Value

a list containing summary statistics and results

Details

Examples

mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ ., mtcars)
summary(fit)

#>   R.Squared.. Adj..R.Squared..    AIC..   RMSE..
#> 1   0.8690158        0.8066423 163.7098 2.146905

#> Loading required package: ggplot2

#> Loading required package: lattice

#> $Metrics
#>   R.Squared.. Adj..R.Squared..    AIC..   RMSE..
#> 1   0.8690158        0.8066423 163.7098 2.146905
#> 
#> $`K-Fold Cross Validation, K=10`
#>   K.Fold.CV.R.squared K.fold.CV.RMSE
#> 1            0.900815       3.227743
#> 
#> $`F-statistic, P-Value`
#> [1] "F(df1: 10, df2: 21) = 13.9325, p < 0"
#> 
#> $`Anova Table (Type III)`
#> Anova Table (Type III tests)
#> 
#> Response: mpg
#>              Sum Sq Df F value  Pr(>F)  
#> (Intercept)   3.035  1  0.4321 0.51812  
#> cyl           0.080  1  0.0114 0.91609  
#> disp          3.917  1  0.5576 0.46349  
#> hp            6.840  1  0.9739 0.33496  
#> drat          1.627  1  0.2317 0.63528  
#> wt           27.014  1  3.8463 0.06325 .
#> qsec          8.864  1  1.2621 0.27394  
#> vs            0.160  1  0.0228 0.88142  
#> am           10.547  1  1.5016 0.23399  
#> gear          1.353  1  0.1926 0.66521  
#> carb          0.407  1  0.0579 0.81218  
#> Residuals   147.494 21                  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> $`Multiple Linear Regression Coefficients`
#>                       B          B*       SE(B)          t    p-value
#> (Intercept) 12.30337416 -0.19294399 18.71788443  0.6573058 0.51812440
#> cyl         -0.11144048 -0.03302235  1.04502336 -0.1066392 0.91608738
#> disp         0.01333524  0.27422706  0.01785750  0.7467585 0.46348865
#> hp          -0.02148212 -0.24438168  0.02176858 -0.9868407 0.33495531
#> drat         0.78711097  0.06982829  1.63537307  0.4813036 0.63527790
#> wt          -3.71530393 -0.60316876  1.89441430 -1.9611887 0.06325215
#> qsec         0.82104075  0.24343220  0.73084480  1.1234133 0.27394127
#> vs           0.31776281  0.05272367  2.10450861  0.1509915 0.88142347
#> am1          2.52022689  0.41815972  2.05665055  1.2254035 0.23398971
#> gear         0.65541302  0.08023404  1.49325996  0.4389142 0.66520643
#> carb        -0.19941925 -0.05344363  0.82875250 -0.2406258 0.81217871
#>