Auto MPG Data Set

City-cyle fuel consumption in miles per gallon, predicted from 3 multivalued discrete variables and 5 continuous variables.

auto_mpg

Format

A data frame with 392 rows and 9 variables:

carname

character; car name.

mpg

numeric; miles per gallon.

cyl

factor; number of cylinders.

disp

numeric; engine displacement.

hp

numeric; horse power.

wt

numeric; weight.

accel

numeric; acceleration.

year

factor; make year.

origin

factor; origin.

Source

ICU Machine Learning Repository. This dataset was taken from the StatLib library which is maintained at Carnegie Mellon University. The dataset was used in the 1983 American Statistical Association Exposition.

Note

Missing values have been removed from the original data set.