Summarise an 'lm' object

Summarize an object of class "lm".

# S3 method for lm
info(x)

Arguments

x

an object of class code"lm"

Value

a 6-component list

fit.indices

includes R-squared, Adjusted R-squared, Akaike's Information Criterion, Root Mean Squared Error, Mean Absolute Error

Ftest

includes F-statistic, Degrees of Freedom, p-value

anova.table

includes a Type-III Analysis of Variance Table (outputted from the Anova function in the car package)

coefficient.table

includes a table of Coefficients, Standardised Coefficients, Standard Errors, t-values, p-values, significance stars

N

sample size

Details

The function info.lm produces a comprehensive summary for a linear model fitted with lm function. In addition to the standard output from the summary.lm function, the output includes the sample size (N), Akaike's Information Criterion (AIC), Mean Absolute Error (MAE), ANOVA table (type III SS), and standardized regression coefficients (B*).

Examples

fit <- lm(mpg ~ hp + wt + accel + origin, data = auto_mpg)
info(fit)
#> MULTIPLE REGRESSION SUMMARY 
#> Model: mpg ~ hp + wt + accel + origin
#> Data : auto_mpg
#> N    : 392
#> 
#> Fit Indices 
#>  R.Squared Adj.R.Squared  AIC RMSE  MAE
#>      0.719         0.716 2238 4.13 3.15
#> 
#> Omnibus Test 
#> F(5,386) = 197.852, p < <2e-16  *** 
#> 
#> Anova Table (type III tests) 
#>               Sum Sq  DF F value Pr(>F)    
#> (Intercept) 5473.655   1 316.038 <0.001 ***
#> hp           220.431   1  12.727 <0.001 ***
#> wt          1017.753   1  58.763 <0.001 ***
#> accel          1.117   1   0.065    0.8    
#> origin       308.474   2   8.905 <0.001 ***
#> Residuals   6685.366 386                   
#> 
#> Regression Coefficients 
#>                    B      B*       SE      t Pr(>|t|)    
#> (Intercept) 43.29141 -0.0928 2.435185 17.777 4.41e-52 ***
#> hp          -0.05643 -0.2783 0.015818 -3.568 4.06e-04 ***
#> wt          -0.00477 -0.5188 0.000622 -7.666 1.46e-13 ***
#> accel       -0.03085 -0.0109 0.121477 -0.254 8.00e-01    
#> origin2      0.97656  0.1251 0.643910  1.517 1.30e-01    
#> origin3      2.75244  0.3527 0.653731  4.210 3.18e-05 ***