Summarise an 'mreg' object

This function serves as a summary method for class 'mreg'. It computes summary statistics and results for a fitted linear model.

# S3 method for mreg
summary(x)

Arguments

x
an object of class 'mreg', presumably an output from the mreg() function.

Value

a list containing summary statistics and results

Details

Examples

fit <- mreg(mpg ~ cyl + hp + wt, mtcars)
summary(fit)

#>   R.Squared.. Adj..R.Squared..    AIC..   RMSE..
#> 1     0.84315        0.8263446 155.4766 2.349338

#> Loading required package: ggplot2

#> Loading required package: lattice

#> $Metrics
#>   R.Squared.. Adj..R.Squared..    AIC..   RMSE..
#> 1     0.84315        0.8263446 155.4766 2.349338
#> 
#> $`K-Fold Cross Validation, K=10`
#>   K.Fold.CV.R.squared K.fold.CV.RMSE
#> 1             0.93574       2.625501
#> 
#> $`F-statistic, P-Value`
#> [1] "F(df1: 3, df2: 28) = 50.1715, p < 0"
#> 
#> $`Anova Table (Type III)`
#> Anova Table (Type III tests)
#> 
#> Response: mpg
#>              Sum Sq Df  F value    Pr(>F)    
#> (Intercept) 2966.77  1 470.3276 < 2.2e-16 ***
#> cyl           18.43  1   2.9213 0.0984801 .  
#> hp            14.55  1   2.3069 0.1400152    
#> wt           115.35  1  18.2873 0.0001995 ***
#> Residuals    176.62 28                       
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> $`Multiple Linear Regression Coefficients`
#>                      B            B*      SE(B)         t      p-value
#> (Intercept) 38.7517874  3.675702e-17 1.78686403 21.687038 4.799399e-19
#> cyl         -0.9416168 -2.790225e-01 0.55091638 -1.709183 9.848010e-02
#> hp          -0.0180381 -2.052024e-01 0.01187625 -1.518838 1.400152e-01
#> wt          -3.1669731 -5.141489e-01 0.74057588 -4.276365 1.994765e-04
#>