Summarise an 'mreg' object

This function serves as a summary method for class "mreg".

# S3 method for mreg
summary(x)

Arguments

x
an object of class "mreg", a result of a call to mreg (which is a wrapper for lm)

Value

The function summary.mreg computes summary statistics and results for a fitted linear model with measures of robustness including:

Fit Indices

includes R-squared, Adjusted R-squared, Akaike's Information Criterion, Root Mean Squared Error

10-fold Cross Validated Fit Indices

includes a more reliable and accurate measure of R-squared, Root Mean Squared Error

Omnibus Test

includes F-statistic, Degrees of Freedom, p-value

Anova Table (Type 3 Tests)

includes a Type-III Analysis of Variance Table (outputted from the Anova() function in the car package)

Regression Coefficients

includes a table of Coefficients, Standardised Coefficients, Standard Errors, t-values, p-values, Significance codes

Examples

mtcars$am <- factor(mtcars$am)
fit <- mreg(mpg ~ hp + wt + am, mtcars)
summary(fit)

#> Loading required package: ggplot2

#> Loading required package: lattice

#> 
#> Multiple Linear Regression Summary
#> ----------------------------------
#> 
#> Formula:mpg ~ hp + wt + am
#> Data   :mtcars
#> 
#> Fit Indices:
#> ------------
#>  R.Squared Adj.R.Squared      AIC     RMSE
#>  0.8398903     0.8227357 156.1348 2.373625
#> 
#> 10-Fold Cross Validated Fit Indices:
#> ------------------------------------
#>  R.squared     RMSE
#>  0.8359868 2.573404
#> 
#> Omnibus Test:
#> -------------
#> F(df1: 3, df2: 28) = 48.96, p < 0  *** 
#> --- 
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 
#> Anova Table (type III tests):
#> -----------------------------
#>              Sum Sq Df  F value    Pr(>F)    
#> (Intercept) 1066.02  1 165.5575 2.824e-13 ***
#> hp            98.03  1  15.2243 0.0005464 ***
#> wt            65.15  1  10.1178 0.0035740 ** 
#> am            14.76  1   2.2918 0.1412682    
#> Residuals    180.29 28                       
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Multiple Linear Regression Coefficients:
#> ----------------------------------------
#>                       B         B*          SE         t      p-value    
#> (Intercept) 34.00287512 -0.1404537 2.642659337 12.866916 2.824030e-13 ***
#> hp          -0.03747873 -0.4263599 0.009605422 -3.901830 5.464023e-04   *
#> wt          -2.87857541 -0.4673283 0.904970538 -3.180850 3.574031e-03   *
#> am1          2.08371013  0.3457322 1.376420152  1.513862 1.412682e-01    
#> --- 
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1